DRYed up sorting of wrestlers in the index for all indexes except related to tag teams.
